delete Entity3D abstract class + clean
It was not a bad idea to have a high level Entity3D class to extend, but for the moment idk exactly the needs of this kind of class, let's thinking about it when we get >10 different entities

Also we move entity crud to an entityManager and frame update in each dedicated class. As the "serialization" will not change (there no way a use other protocol than TLV data imo) we can consider to put the frame parsing directly in needed part of the pipeline.
Here:
1. ws parses familyId
2. entityManager parses id
3. Puwu3D parses all 3D data to display